Turn off CPACK_NSIS_MODIFY_PATH.
Turning this on inserts a page in the installer which gives the user the
option of adding LÖVE to PATH, except it doesn't really work, even if the
user opts that it should happen.

In that case, it's better to not give the user a choice at all, and disable
the whole thing.